Добавлено маршрутизацію для сторінок Home, Login та Register; оновлено API URL для пошуку чатів; видалено старі компоненти Login та Register.
