+++ src/App.jsx
@@ -1,12 +1,17 @@
 import "./App.css";
-import ChatArea from "./components/Chat/ChatArea";
-import Sidebar from "./components/SideBar/Sidebar";
+import Home from "./pages/Home";
+import Register from "./pages/Register";
+import Login from "./pages/Login";
+import { Routes, Route } from "react-router-dom";
 
 function App() {
   return (
-    <div className="app">
-      <Sidebar />
-      {/* <ChatArea /> */}
+    <div>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/register" element={<Register />} />
+        <Route path="/login" element={<Login />} />
+      </Routes>
     </div>
   );
 }
